LPD-22028 Merge export and import tests into one class
